requirements.txt
openai
langchain
numpy
usearch
pystemmer
bm25s
pymupdf
instructor
Flask